move checker and transformer to different cpp files
