+++ test-data/packet.yaml
@@ -1,6 +1,35 @@
 # Copyright (c) Prevail Verifier contributors.
 # SPDX-License-Identifier: MIT
 ---
+test-case: read 8bit is unsigned
+
+pre: ["meta_offset=0", "packet_size=20",
+      "r1.type=packet", "r1.packet_offset=0", "r1.svalue=[4098, 2147418112]", "r1.uvalue=[4098, 2147418112]"]
+
+code:
+  <start>: |
+    r2 = *(u8 *)(r1 + 0)
+
+post: ["meta_offset=0", "packet_size=20",
+       "r1.type=packet", "r1.packet_offset=0", "r1.svalue=[4098, 2147418112]", "r1.uvalue=[4098, 2147418112]",
+       "r2.type=number", "r2.svalue=[0, 255]", "r2.uvalue=[0, 255]"]
+messages: []
+---
+test-case: read 16bit is unsigned
+
+pre: ["meta_offset=0", "packet_size=20",
+      "r1.type=packet", "r1.packet_offset=0", "r1.svalue=[4098, 2147418112]", "r1.uvalue=[4098, 2147418112]"]
+
+code:
+  <start>: |
+    r2 = *(u16 *)(r1 + 0)
+
+post: ["meta_offset=0", "packet_size=20",
+       "r1.type=packet", "r1.packet_offset=0", "r1.svalue=[4098, 2147418112]", "r1.uvalue=[4098, 2147418112]",
+       "r2.type=number", "r2.svalue=[0, 65535]", "r2.uvalue=[0, 65535]"]
+messages: []
+
+---
 test-case: simple invalid write
 
 pre: ["meta_offset=0",
